requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
app==0.0.1
bcrypt==4.2.1
fastapi==0.115.5
greenlet==3.1.1
idna==3.10
passlib==1.7.4
pydantic==2.10.2
pydantic_core==2.27.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
psycopg2-binary==2.9.10
python-dotenv==1.0.0
uvicorn==0.22.0
email-validator==2.0.0
PyJWT==2.8.0
